IE11 Simplelayout overlay fix.
Since we place save / cancel buttons absolutely in the overlay we need
to make sure that no content hides beneath the form buttons.

This was done with a padding, but unfortunately this is not supported by
IE 11. In order to support IE 11 we do the same with an ``:after``
pseudo element.