Define minimal styling for portal language selector.
The portal language selector will be reworked with ftw.mobile.
This solution is a quick fix.
The selector has been touched in
https://github.com/4teamwork/plonetheme.blueberry/commit/61a2f247b9b1cdb8b1e0b9587c881ede99d62819
but we don't know exactly what the effects should be there.
The language selector is only in use at http://sicher-ses.ch/de.
So when http://sicher-ses.ch/de will be updated before the rework
we should keep an eye on it.