Force all columns to use the full page in print mode.
In print mode the columns should use the full page because chrome
reacts on screen media queries in print mode as well. So force all columns
to use the full page width.

Additionally define the size of the printed page to be a DIN A4. Also make these
restrictions on the HTML and body tag.